# docs/openapi/admin-api.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Mistvale Admin API",
    "version": "0.1.0-ea.0",
    "description": "The Admin API the Mistvale Admin Suite consumes, plus the public content bundle it previews. Generated from the Zod contracts in @mistvale/shared — edit those, then run `pnpm openapi`."
  },
  "servers": [
    {
      "url": "https://play.pathlands.cc",
      "description": "Production"
    }
  ],
...
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "AUTH_REQUIRED",
              "FORBIDDEN",
              "VALIDATION",
              "INVALID_CREDENTIALS",
              "ALREADY_EXISTS",
              "NOT_FOUND",
              "INSUFFICIENT_FUNDS",
              "ENERGY_LOW",
              "ROSTER_FULL",
              "COOLDOWN",
              "LOCKED_CONTENT",
              "IDEMPOTENT_REPLAY",
              "RATE_LIMITED",
              "CONTENT_STALE",
              "ACCOUNT_BANNED",
              "PASSWORD_CHANGE_REQUIRED",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {},
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "additionalProperties": false
      },
      "AssetDef": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 9999
          },
          "kind": {
            "type": "string",
            "enum": [
              "unit",
              "vfx",
              "ui",
              "audio"
            ]
          },
          "source": {
            "default": "repo",
            "type": "string",
            "enum": [
              "repo",
              "upload"
            ]
          },
          "basePath": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "tracks": {
            "default": {},
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "object",
              "properties": {
                "frames": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 64
                },
                "fps": {
                  "default": 9,
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 30
                },
                "loop": {
                  "default": false,
                  "type": "boolean"
                }
              },
              "required": [
                "frames",
                "fps",
                "loop"
              ],
              "additionalProperties": false
            }
          },
          "stillPath": {
            "default": "",
            "type": "string",
            "maxLength": 200
          },
          "avatarPath": {
            "default": "",
            "type": "string",
            "maxLength": 200
          },
          "tint": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$"
          }
        },
        "required": [
          "key",
          "sortOrder",
          "kind",
          "source",
          "basePath",
          "tracks",
          "stillPath",
          "avatarPath"
        ],
        "additionalProperties": false
      },
...
      "ContentBundle": {
        "type": "object",
        "properties": {
          "rev": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "publishedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "factions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FactionDef"
            }
          },
          "statuses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StatusDef"
            }
          },
          "skills": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SkillDef"
            }
          },
          "assets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetDef"
            }
          },
          "champions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ChampionDef"
            }
          },
          "enemies": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EnemyDef"
            }
          },
          "gearSets": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GearSetDef"
            }
          },
          "gearSlots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GearSlotDef"
            }
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ItemDef"
            }
          },
          "campaignChapters": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CampaignChapterDef"
            }
          },
          "stages": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/StageDef"
            }
          },
          "config": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string"
                },
                {
                  "type": "boolean"
                },
                {
                  "type": "object",
                  "propertyNames": {
                    "type": "string"
                  },
                  "additionalProperties": {}
                },
                {
                  "type": "array",
                  "items": {}
                }
              ]
            }
          }
        },
        "required": [
          "rev",
          "publishedAt",
          "factions",
          "statuses",
          "skills",
          "assets",
          "champions",
          "enemies",
          "gearSets",
          "gearSlots",
          "items",
          "campaignChapters",
          "stages",
          "config"
        ],
        "additionalProperties": false
      },
...
      "ContentEntry": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string"
          },
          "data": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "state": {
            "type": "string",
            "enum": [
              "live",
              "draft",
              "deleting"
            ]
          },
          "updatedAt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          },
          "updatedBy": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "key",
          "data",
          "state",
          "updatedAt",
          "updatedBy"
        ],
        "additionalProperties": false
      },
...
      "FactionDef": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 9999
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 48
          },
          "lore": {
            "default": "",
            "type": "string",
            "maxLength": 2000
          },
          "icon": {
            "default": "",
            "type": "string",
            "maxLength": 64
          }
        },
        "required": [
          "key",
          "sortOrder",
          "name",
          "lore",
          "icon"
        ],
        "additionalProperties": false
      },
...
      "GearSetDef": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 9999
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 48
          },
          "lore": {
            "default": "",
            "type": "string",
            "maxLength": 600
          },
          "pieces": {
            "anyOf": [
              {
                "type": "number",
                "const": 2
              },
              {
                "type": "number",
                "const": 4
              }
            ]
          },
          "bonusType": {
            "type": "string",
            "enum": [
              "stat",
              "lifesteal",
              "regen",
              "provokeOnHit",
              "stunOnHit",
              "burnOnHit",
              "counterOnHit",
              "tmOnDamageTaken"
            ]
          },
          "bonus": {
            "type": "object",
            "properties": {
              "stat": {
                "type": "string",
                "enum": [
                  "hp",
                  "atk",
                  "def",
                  "spd",
                  "critRate",
                  "critDmg",
                  "res",
                  "acc"
                ]
              },
              "pct": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              },
              "flat": {
                "type": "number",
                "minimum": 0,
                "maximum": 200
              },
              "chance": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              },
              "turns": {
                "type": "integer",
                "minimum": 1,
                "maximum": 4
              }
            },
            "additionalProperties": false
          }
        },
        "required": [
          "key",
          "sortOrder",
          "name",
          "lore",
          "pieces",
          "bonusType",
          "bonus"
        ],
        "additionalProperties": false
      },
      "GearSlotDef": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "enum": [
              "weapon",
              "helm",
              "shield",
              "gauntlets",
              "cuirass",
              "boots",
              "ring",
              "amulet",
              "banner"
            ]
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 32
          },
          "allowedMainStats": {
            "minItems": 1,
            "type": "array",
            "items": {
              "type": "string",
              "enum": [
                "hp",
                "atk",
                "def",
                "spd",
                "critRate",
                "critDmg",
                "res",
                "acc"
              ]
            }
          },
          "allowsPercentMain": {
            "default": false,
            "type": "boolean"
          },
          "accessory": {
            "default": false,
            "type": "boolean"
          },
          "ascensionRequired": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 6
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 99
          }
        },
        "required": [
          "key",
          "name",
          "allowedMainStats",
          "allowsPercentMain",
          "accessory",
          "ascensionRequired",
          "sortOrder"
        ],
        "additionalProperties": false
      },
...
      "ListContentEntriesResponse": {
        "type": "object",
        "properties": {
          "contentType": {
            "type": "string",
            "enum": [
              "faction",
              "status",
              "skill",
              "asset",
              "champion",
              "enemy",
              "gearSet",
              "gearSlot",
              "item",
              "campaignChapter",
              "stage",
              "gameConfig"
            ]
          },
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ContentEntry"
            }
          }
        },
        "required": [
          "contentType",
          "items"
        ],
        "additionalProperties": false
      },
...
      "StatusDef": {
        "type": "object",
        "properties": {
          "key": {
            "type": "string",
            "minLength": 2,
            "maxLength": 64,
            "pattern": "^[a-z][a-z0-9_]*$"
          },
          "sortOrder": {
            "default": 0,
            "type": "integer",
            "minimum": 0,
            "maximum": 9999
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 48
          },
          "kind": {
            "type": "string",
            "enum": [
              "buff",
              "debuff"
            ]
          },
          "engineType": {
            "type": "string",
            "enum": [
              "statModifier",
              "damageOverTime",
              "healOverTime",
              "shield",
              "skipTurn",
              "skipTurnBreakOnDamage",
              "forceTargetA1",
              "blockBuffs",
              "blockDebuffs",
              "counterattack",
              "allyProtection",
              "reflectDamage",
              "lifesteal",
              "healReduction",
              "unkillable"
            ]
          },
          "family": {
            "type": "string",
            "minLength": 1,
            "maxLength": 48
          },
          "potency": {
            "default": 1,
            "type": "integer",
            "minimum": 1,
            "maximum": 10
          },
          "params": {
            "default": {
              "tick": "none",
              "maxStacks": 1
            },
            "type": "object",
            "properties": {
              "stat": {
                "type": "string",
                "enum": [
                  "hp",
                  "atk",
                  "def",
                  "spd",
                  "critRate",
                  "critDmg",
                  "res",
                  "acc"
                ]
              },
              "pct": {
                "type": "number",
                "minimum": -100,
                "maximum": 200
              },
              "flat": {
                "type": "number",
                "minimum": -200,
                "maximum": 200
              },
              "tickPct": {
                "type": "number",
                "minimum": 0,
                "maximum": 50
              },
              "tick": {
                "default": "none",
                "type": "string",
                "enum": [
                  "none",
                  "ownerTurnStart",
                  "ownerTurnEnd"
                ]
              },
              "maxStacks": {
                "default": 1,
                "type": "integer",
                "minimum": 1,
                "maximum": 10
              },
              "ratio": {
                "type": "number",
                "minimum": 0,
                "maximum": 100
              }
            },
            "required": [
              "tick",
              "maxStacks"
            ],
            "additionalProperties": false
          },
          "icon": {
            "default": "",
            "type": "string",
            "maxLength": 64
          },
          "description": {
            "default": "",
            "type": "string",
            "maxLength": 400
          }
        },
        "required": [
          "key",
          "sortOrder",
          "name",
          "kind",
          "engineType",
          "family",
          "potency",
          "params",
          "icon",
          "description"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "mv_session",
        "description": "Session cookie set by the login endpoint."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Session token, for clients that cannot use cookies."
      }
    }
  },
  "paths": {
    "/api/content": {
      "get": {
        "operationId": "getContentBundle",
        "summary": "Fetch the live content bundle",
        "tags": [
          "content"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ContentBundle"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "304": {
            "description": "Content has not changed since the ETag you sent."
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": false
                    },
                    "error": {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "description": "Served from memory and cached by revision. Send `If-None-Match` with the ETag from a previous response to get a 304 when content has not moved.",
        "parameters": [
          {
            "name": "If-None-Match",
            "in": "header",
            "required": false,
            "description": "ETag from a previous bundle response.",
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/admin/api/content/{type}": {
      "get": {
        "operationId": "listContentEntries",
        "summary": "Every entity of one content type",
        "tags": [
          "admin"
        ],
        "responses": {
          "200": {
            "description": "Success.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ListContentEntriesResponse"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "data",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "No session, or the credentials were wrong.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": false
                    },
                    "error": {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "403": {
            "description": "The session is valid but the account lacks the admin rank.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": false
                    },
                    "error": {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "404": {
            "description": "No such content type, entity or revision.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": false
                    },
                    "error": {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "500": {
            "description": "Unexpected server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "const": false
                    },
                    "error": {
                      "$ref": "#/components/schemas/ApiError"
                    },
                    "rev": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "ok",
                    "error",
                    "rev"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "type",
            "in": "path",
            "required": true,
            "description": "Content type path segment.",
            "schema": {
              "type": "string",
              "enum": [
                "factions",
                "statuses",
                "skills",
                "assets",
                "champions",
                "enemies",
                "gear-sets",
                "gear-slots",
                "items",
                "chapters",
                "stages",
                "config"
              ]
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  }
}